Use apt's CommandLine parser
Previously aptitude was using it's own command line
parser implemented on top of getopt.  Replacing this
with the standard apt class means there are no longer
subtle quirks with our parser (such as "-qq" not
being valid, when it is fine for other apt programs).

The parser is closely tied to the Configuration, and
reads options directly in to the global apt config.

Some options were tracked independently of the global
config; this change is the first part rectifying that
situation.  The goal is to keep all program options,
configuration state, etc. in the global config rather
than passing state variables between functions.  Many
options which previously did not have an associated
configuration item now do; some remain undocumented
for now as they are transitional.

* handle "-qq" like other apt-utils
* properly process all -o command line options (Closes: #587671)
* unknown command line options trigger an error (Closes: #434502)
* --add-user-tag-to, --remove-user-tag-from are working;
  no bug reports suggests that noone has noticed they
  were broken

New aliases for some command-line options:
  --yes, for --assume-yes (apt-get)
  --default-release, for --target-release (apt-utils)
  --install-on-startup, for -i
  --update-on-startup, for -u

New configuration item:
  Aptitude::CmdLine::Sorting, equivalent to --sort

Other new configuration items are for internal use only
at this point.

Closes: #587671
Closes: #434502