Add pkgset helpers
Ported some code from apt cacheset.cc to populate a pkgset.
The function pkgset_from_string will fill a given pkgset with
either the exactly named package, or the packages matched by a
string predicated by is_pattern.

In the future, we should look to directly use the apt cacheset
infrastructure.  This is currently blocked by our custom
implementation of pkgCacheFile (see aptcache.h).

Errors that occur in looking up the package (such as no package
with that name, or an invalid search pattern) are pushed to
the global error list with a customizable error_type (default:
ERROR).