Do not do FixMissing automatically for command-line actions
When a package is not available or fails to download
incorrectly aptitude would try to continue without it.
This made it hard to detect errors automatically using
the exit status.

This behaviour is no longer the default, instead if there
are missing packages an error will be displayed and the
program exit with failure (100).  If the user wishes to
continue anyway, they can use --fix-missing or configure
Aptitude::CmdLine::Fix-Missing in apt.conf.

  * [cmdline]: An install run will no longer proceed if any
               package was unavailable or failed to download
               correctly, instead an error will be shown
               and exit with non-zero status (100).
               (Closes: #121313, #302784)

               The old behaviour can be obtained with the new
               --fix-missing option or setting
               Aptitude::CmdLine::Fix-Missing.

Closes: #121313
Closes: #302784