apt: Derive aptitudeCacheFile from pkgCacheFile (apt-pkg)
In ancient times it was not possible to have aptitudes CacheFile class
inherit from that in apt-pkg.  This inhibits using directly many
apt-pkg interfaces, resulting in unnecessary duplication of code and
implementation of less effective or standardized solutions.

The blocking issues are resolved for some time.  Mainly that is
apt-pkg including features that were once aptitude specific and
refinement of apt-pkg to better support needs of clients.

With aptitudeCacheFile subclassing pkgCacheFile, work can begin to
remove much of the remaining duplication and integrate more of the
standard apt-pkg interfaces, such as CacheSets.

* src/generic/apt/aptcache.cc:
  - Derive aptitudeCacheFile from pkgCacheFile.
  - Destructor does not need to ::ReleaseLock as ~pkgCacheFile does
    that.
  - Override ::BuildDepCache to construct an aptitudeDepCache
    instance.
  - Refactor ::Open using ::Build*.
  - Set pkgCacheFile::Generate "true" to maintain aptitude's current
    behaviour; may want to remove this in the future and also change
    download_update_manager.
  - Take ownership of the global apt_source_list.